move types to shared file
